refactor(harness): own resource discovery parsing and filesystem
